requirements.txt
click == 7.1.2
Flask==1.1.2
werkzeug==1.0.1
psycopg2-binary == 2.8.6
pytz == 2021.1
